Crate polkavm

Structs

CallArgs

A helper struct used when calling into a function exported by the guest program.

Caller

A handle used to access the execution context.

CallerRef

A handle used to access the execution context, with erased lifetimes for convenience.

Config

Engine

Error

ExportIndex

The index of an exported function to be called.

Gas

A gas value used for gas metering.

Instance

InstancePre

Linker

MemoryMap

The memory map of a given guest program.

Module

A compiled PolkaVM program module.

ModuleConfig

The configuration for a module.

ProgramBlob

A partially deserialized PolkaVM program.

ProgramParseError

StateArgs

Trap

Enums

BackendKind

ExecutionError

GasMeteringKind

The type of gas metering.

Reg

SandboxKind

Traits

AsUninitSliceMut