polkavm

Struct ProgramBlob

pub struct ProgramBlob { /* private fields */ }

A partially deserialized PolkaVM program.

Implementations

impl ProgramBlob

pub fn blob_length(raw_blob: &[u8]) -> Option<u64>

Parses the blob length information from the given raw_blob bytes.

Returns None if raw_blob doesnโt contain enough bytes to read the length.

pub fn parse(bytes: ArcBytes) -> Result<ProgramBlob, ProgramParseError>

Parses the given bytes into a program blob.

pub fn from_parts(parts: ProgramParts) -> Result<ProgramBlob, ProgramParseError>

Creates a program blob from parts.

pub fn unique_id(&self) -> u64

Returns an unique ID of the program blob.

This is an automatically incremented counter every time a ProgramBlob is created.

pub fn isa(&self) -> InstructionSetKind

Returns the instruction set of this program blob.

pub fn is_64_bit(&self) -> bool

Returns whether the blob contains a 64-bit program.

pub fn unique_hash(&self, include_debug: bool) -> Hash

Calculates an unique hash of the program blob.

pub fn ro_data(&self) -> &[u8]

Returns the contents of the read-only data section.

This only covers the initial non-zero portion of the section; use ro_data_size to get the full size.

pub fn ro_data_size(&self) -> u32

Returns the size of the read-only data section.

This can be larger than the length of ro_data, in which case the rest of the space is assumed to be filled with zeros.

pub fn rw_data(&self) -> &[u8]

Returns the contents of the read-write data section.

This only covers the initial non-zero portion of the section; use rw_data_size to get the full size.

pub fn rw_data_size(&self) -> u32

Returns the size of the read-write data section.

This can be larger than the length of rw_data, in which case the rest of the space is assumed to be filled with zeros.

pub fn stack_size(&self) -> u32

Returns the initial size of the stack.

pub fn code(&self) -> &[u8]

Returns the program code in its raw form.

pub fn bitmask(&self) -> &[u8]

Returns the code bitmask in its raw form.

pub fn imports(&self) -> Imports<'_>

pub fn exports(&self) -> impl Iterator<Item = ProgramExport<&[u8]>> + Clone

Returns an iterator over program exports.

pub fn visit<T>( &self, dispatch_table: T, visitor: &mut <T as OpcodeVisitor>::State, )
where T: OpcodeVisitor<ReturnTy = ()>,

Visits every instruction in the program.

pub fn instructions(&self) -> Instructions<'_, InstructionSetKind>

pub fn instructions_bounded_at( &self, offset: ProgramCounter, ) -> Instructions<'_, InstructionSetKind>

pub fn instructions_with_isa<I>( &self, instruction_set: I, ) -> Instructions<'_, I>
where I: InstructionSet,

Returns an iterator over all of the instructions in the program.

WARNING: this is unbounded and has O(n) complexity; just creating this iterator can iterate over the whole program, even if next is never called!

pub fn instructions_bounded_at_with_isa<I>( &self, instruction_set: I, offset: ProgramCounter, ) -> Instructions<'_, I>
where I: InstructionSet,

Returns an interator over instructions starting at a given offset.

This iterator is bounded and has O(1) complexity.

pub fn is_jump_target_valid<I>( &self, instruction_set: I, target: ProgramCounter, ) -> bool
where I: InstructionSet,

Returns whether the given program counter is a valid target for a jump.

pub fn jump_table(&self) -> JumpTable<'_>

Returns a jump table.

pub fn get_debug_string(&self, offset: u32) -> Result<&str, ProgramParseError>

Returns the debug string for the given relative offset.

pub fn get_debug_line_program_at( &self, program_counter: ProgramCounter, ) -> Result<Option<LineProgram<'_>>, ProgramParseError>

Returns the line program for the given instruction.

pub fn estimate_interpreter_memory_usage( &self, args: EstimateInterpreterMemoryUsageArgs, ) -> Result<ProgramMemoryInfo, &'static str>