Struct polkavm::Config

pub struct Config { /* private fields */ }

Implementations

impl Config

pub fn new() -> Self

Creates a new default configuration.

pub fn from_env() -> Result<Self, Error>

Creates a new default configuration and seeds it from the environment variables.

pub fn set_backend(&mut self, backend: Option<BackendKind>) -> &mut Self

Forces the use of a given backend.

Default: None (automatically pick the best available backend)

Corresponding environment variable: POLKAVM_BACKEND (auto, compiler, interpreter)

pub fn backend(&self) -> Option<BackendKind>

Gets the currently set backend, if any.

pub fn set_sandbox(&mut self, sandbox: Option<SandboxKind>) -> &mut Self

Forces the use of a given sandbox.

Default: None (automatically pick the best available sandbox)

Corresponding environment variable: POLKAVM_SANDBOX (auto, linux, generic)

pub fn sandbox(&self) -> Option<SandboxKind>

Gets the currently set sandbox, if any.

pub fn set_crosscheck(&mut self, value: bool) -> &mut Self

Enables execution cross-checking.

This will run an interpreter alongside the recompiler and cross-check their execution.

Should only be used for debugging purposes and never enabled by default in production.

Default: false

Corresponding environment variable: POLKAVM_CROSSCHECK (false, true)

pub fn crosscheck(&self) -> bool

Returns whether cross-checking is enabled.

pub fn set_allow_experimental(&mut self, value: bool) -> &mut Self

Enabling this makes it possible to enable other experimental settings which are not meant for general use and can introduce unsafety, break determinism, or just simply be totally broken.

This should NEVER be used in production unless you know what youโre doing.

Default: false

Corresponding environment variable: POLKAVM_ALLOW_EXPERIMENTAL (true, false)

pub fn set_worker_count(&mut self, value: usize) -> &mut Self

Sets the number of worker sandboxes that will be permanently kept alive by the engine.

This doesnโt limit the number of instances that can be instantiated at the same time; it will just tell the engine how many sandboxes should be cached between instantiations.

For the Linux sandbox this will decide how many worker processes are kept alive.

This only has an effect when using a recompiler. For the interpreter this setting will be ignored.

Default: 2

Corresponding environment variable: POLKAVM_WORKER_COUNT

pub fn worker_count(&self) -> usize

Returns the number of worker sandboxes that will be permanently kept alive by the engine.

pub fn dynamic_paging(&self) -> bool

Returns whether dynamic paging is enabled.

pub fn set_dynamic_paging(&mut self, value: bool) -> &mut Self

Sets whether dynamic paging is enabled.

Default: false